refactor: core:ui 모듈 내 유틸리티 및 상태 관리 파일 구조 개선
core:ui 모듈의 유지보수성을 높이기 위해 파일 위치를 기능별 패키지(`util`, `state`)로 재배치하고 관련 임포트를 업데이트했습니다.

* **refactor: 유틸리티 및 상태 관련 파일 패키지 이동**
    * `AdvancedImePadding.kt` 및 `OnHeightChanged.kt` 파일을 `com.team.prezel.core.ui.util` 패키지로 이동했습니다.
    * `LocalSnackbarHostState.kt` 파일을 `com.team.prezel.core.ui.state` 패키지로 이동했습니다.

* **style: 변경된 구조에 따른 임포트 경로 수정**
    * `ProfileScreen`, `HistoryScreen`, `HomeScreen`, `LoginScreen` 등 각 피처 모듈에서 변경된 패키지 경로를 반영하도록 수정했습니다.
    * `PrezelApp` 및 `DoubleBackToExitHandler` 내 `LocalSnackbarHostState` 참조 경로를 업데이트했습니다.
